backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiomysql>=0.2.0
pydantic>=2.6.0
python-slugify>=8.0.0
python-dotenv>=1.0.0
sqlparse>=0.5.0
pytest>=8.0.0
httpx>=0.27.0
